Try upload image to backend
I create a new activity named 'ConfigureActivity', the path to open it is click the '设置' button in 'Personal Fragment'.
And in this activity, you can click the button here to select Your new Touxiang image from the system album. Since the backend hasn't complete the function, I don't know if the complishment is right or not.